+++ NEWS.md
@@ -1,5 +1,7 @@
 # dsn 0.0.1
 
+* New functions for 'datatype()' . 
+
 * New `mem()` function to reference an in memory array. 
 
 * New functions `sds()` and `vrtcon()`.